@@ -1,50 +1,51 @@
 {
-  "name": "qs",
-  "description": "A querystring parser that supports nesting and arrays, with a depth limit",
-  "homepage": "https://github.com/ljharb/qs",
-  "version": "6.3.2",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ljharb/qs.git"
-  },
-  "main": "lib/index.js",
-  "contributors": [
-    {
-      "name": "Jordan Harband",
-      "email": "ljharb@gmail.com",
-      "url": "http://ljharb.codes"
-    }
-  ],
-  "keywords": [
-    "querystring",
-    "qs"
-  ],
-  "engines": {
-    "node": ">=0.6"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@ljharb/eslint-config": "^11.0.0",
-    "browserify": "^14.1.0",
-    "covert": "^1.1.0",
-    "eslint": "^3.17.0",
-    "evalmd": "^0.0.17",
-    "iconv-lite": "^0.4.15",
-    "mkdirp": "^0.5.1",
-    "parallelshell": "^2.0.0",
-    "qs-iconv": "^1.0.4",
-    "safe-publish-latest": "^1.1.1",
-    "tape": "^4.6.3"
-  },
-  "scripts": {
-    "prepublish": "safe-publish-latest && npm run dist",
-    "pretest": "npm run --silent readme && npm run --silent lint",
-    "test": "npm run --silent coverage",
-    "tests-only": "node test",
-    "readme": "evalmd README.md",
-    "lint": "eslint lib/*.js test/*.js",
-    "coverage": "covert test",
-    "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js"
-  },
-  "license": "BSD-3-Clause"
+    "name": "qs",
+    "description": "A querystring parser that supports nesting and arrays, with a depth limit",
+    "homepage": "https://github.com/ljharb/qs",
+    "version": "6.5.1",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/ljharb/qs.git"
+    },
+    "main": "lib/index.js",
+    "contributors": [
+        {
+            "name": "Jordan Harband",
+            "email": "ljharb@gmail.com",
+            "url": "http://ljharb.codes"
+        }
+    ],
+    "keywords": [
+        "querystring",
+        "qs"
+    ],
+    "engines": {
+        "node": ">=0.6"
+    },
+    "dependencies": {},
+    "devDependencies": {
+        "@ljharb/eslint-config": "^12.2.1",
+        "browserify": "^14.4.0",
+        "covert": "^1.1.0",
+        "editorconfig-tools": "^0.1.1",
+        "eslint": "^4.6.1",
+        "evalmd": "^0.0.17",
+        "iconv-lite": "^0.4.18",
+        "mkdirp": "^0.5.1",
+        "qs-iconv": "^1.0.4",
+        "safe-publish-latest": "^1.1.1",
+        "tape": "^4.8.0"
+    },
+    "scripts": {
+        "prepublish": "safe-publish-latest && npm run dist",
+        "pretest": "npm run --silent readme && npm run --silent lint",
+        "test": "npm run --silent coverage",
+        "tests-only": "node test",
+        "readme": "evalmd README.md",
+        "prelint": "editorconfig-tools check * lib/* test/*",
+        "lint": "eslint lib/*.js test/*.js",
+        "coverage": "covert test",
+        "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js"
+    },
+    "license": "BSD-3-Clause"
 }
