@@ -2020,7 +2020,7 @@ module.exports = __webpack_require__(29);
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-Object.defineProperty(exports,"__esModule",{value:true});exports.default=void 0;var _react=_interopRequireWildcard(__webpack_require__(1));var _reactRouterDom=__webpack_require__(32);var _multipleExports=__webpack_require__(61);function _interopRequireWildcard(obj){if(obj&&obj.__esModule){return obj;}else{var newObj={};if(obj!=null){for(var key in obj){if(Object.prototype.hasOwnProperty.call(obj,key)){var desc=Object.defineProperty&&Object.getOwnPropertyDescriptor?Object.getOwnPropertyDescriptor(obj,key):{};if(desc.get||desc.set){Object.defineProperty(newObj,key,desc);}else{newObj[key]=obj[key];}}}}newObj.default=obj;return newObj;}}function _extends(){_extends=Object.assign||function(target){for(var i=1;i<arguments.length;i++){var source=arguments[i];for(var key in source){if(Object.prototype.hasOwnProperty.call(source,key)){target[key]=source[key];}}}return target;};return _extends.apply(this,arguments);}function _typeof(obj){if(typeof Symbol==="function"&&typeof Symbol.iterator==="symbol"){_typeof=function _typeof(obj){return typeof obj;};}else{_typeof=function _typeof(obj){return obj&&typeof Symbol==="function"&&obj.constructor===Symbol&&obj!==Symbol.prototype?"symbol":typeof obj;};}return _typeof(obj);}function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor)){throw new TypeError("Cannot call a class as a function");}}function _defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||false;descriptor.configurable=true;if("value"in descriptor)descriptor.writable=true;Object.defineProperty(target,descriptor.key,descriptor);}}function _createClass(Constructor,protoProps,staticProps){if(protoProps)_defineProperties(Constructor.prototype,protoProps);if(staticProps)_defineProperties(Constructor,staticProps);return Constructor;}function _possibleConstructorReturn(self,call){if(call&&(_typeof(call)==="object"||typeof call==="function")){return call;}if(!self){throw new ReferenceError("this hasn't been initialised - super() hasn't been called");}return self;}function _inherits(subClass,superClass){if(typeof superClass!=="function"&&superClass!==null){throw new TypeError("Super expression must either be null or a function");}subClass.prototype=Object.create(superClass&&superClass.prototype,{constructor:{value:subClass,enumerable:false,writable:true,configurable:true}});if(superClass)Object.setPrototypeOf?Object.setPrototypeOf(subClass,superClass):subClass.__proto__=superClass;}var Page1=/*#__PURE__*/function(_React$Component){_inherits(Page1,_React$Component);function Page1(){_classCallCheck(this,Page1);return _possibleConstructorReturn(this,(Page1.__proto__||Object.getPrototypeOf(Page1)).apply(this,arguments));}_createClass(Page1,[{key:"render",value:function render(){return _react.default.createElement("div",null,_react.default.createElement(_multipleExports.A,null));}}]);return Page1;}(_react.default.Component);var Page3=/*#__PURE__*/function(_React$Component2){_inherits(Page3,_React$Component2);function Page3(){_classCallCheck(this,Page3);return _possibleConstructorReturn(this,(Page3.__proto__||Object.getPrototypeOf(Page3)).apply(this,arguments));}_createClass(Page3,[{key:"render",value:function render(){return _react.default.createElement("div",null);}}]);return Page3;}(_react.default.Component);var Page2=/*#__PURE__*/function(_React$Component3){_inherits(Page2,_React$Component3);function Page2(){_classCallCheck(this,Page2);return _possibleConstructorReturn(this,(Page2.__proto__||Object.getPrototypeOf(Page2)).apply(this,arguments));}_createClass(Page2,[{key:"render",value:function render(){return _react.default.createElement("div",null);}}]);return Page2;}(_react.default.Component);var Dash=function Dash(component){return function(routeProps){var DashElement=(0,_react.createElement)(component,_extends({resource:resource},routeProps));return(0,_react.createElement)("div",{},(0,_react.createElement)(Sidebar,_extends({resource:resource},routeProps)),(0,_react.createElement)(component,_extends({resource:resource},routeProps)));};};var Routes=/*#__PURE__*/function(_React$Component4){_inherits(Routes,_React$Component4);function Routes(){_classCallCheck(this,Routes);return _possibleConstructorReturn(this,(Routes.__proto__||Object.getPrototypeOf(Routes)).apply(this,arguments));}_createClass(Routes,[{key:"render",value:function render(){var listReasource={};return _react.default.createElement(_reactRouterDom.BrowserRouter,null,_react.default.createElement(_reactRouterDom.Switch,null,_react.default.createElement(_reactRouterDom.Route,{exact:true,path:"/login",component:Page1}),_react.default.createElement(_reactRouterDom.Route,{exact:true,path:"/user-lise",component:Page2}),_react.default.createElement(_reactRouterDom.Route,{exact:true,path:"/user/:firstName/:lastName/:id",component:Page3})));}}]);return Routes;}(_react.default.Component);exports.default=Routes;
+Object.defineProperty(exports,"__esModule",{value:true});exports.default=void 0;var _react=_interopRequireWildcard(__webpack_require__(1));var _reactRouterDom=__webpack_require__(32);var _multipleExports=__webpack_require__(61);function _interopRequireWildcard(obj){if(obj&&obj.__esModule){return obj;}else{var newObj={};if(obj!=null){for(var key in obj){if(Object.prototype.hasOwnProperty.call(obj,key)){var desc=Object.defineProperty&&Object.getOwnPropertyDescriptor?Object.getOwnPropertyDescriptor(obj,key):{};if(desc.get||desc.set){Object.defineProperty(newObj,key,desc);}else{newObj[key]=obj[key];}}}}newObj.default=obj;return newObj;}}function _extends(){_extends=Object.assign||function(target){for(var i=1;i<arguments.length;i++){var source=arguments[i];for(var key in source){if(Object.prototype.hasOwnProperty.call(source,key)){target[key]=source[key];}}}return target;};return _extends.apply(this,arguments);}function _typeof(obj){if(typeof Symbol==="function"&&typeof Symbol.iterator==="symbol"){_typeof=function _typeof(obj){return typeof obj;};}else{_typeof=function _typeof(obj){return obj&&typeof Symbol==="function"&&obj.constructor===Symbol&&obj!==Symbol.prototype?"symbol":typeof obj;};}return _typeof(obj);}function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor)){throw new TypeError("Cannot call a class as a function");}}function _defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||false;descriptor.configurable=true;if("value"in descriptor)descriptor.writable=true;Object.defineProperty(target,descriptor.key,descriptor);}}function _createClass(Constructor,protoProps,staticProps){if(protoProps)_defineProperties(Constructor.prototype,protoProps);if(staticProps)_defineProperties(Constructor,staticProps);return Constructor;}function _possibleConstructorReturn(self,call){if(call&&(_typeof(call)==="object"||typeof call==="function")){return call;}if(!self){throw new ReferenceError("this hasn't been initialised - super() hasn't been called");}return self;}function _inherits(subClass,superClass){if(typeof superClass!=="function"&&superClass!==null){throw new TypeError("Super expression must either be null or a function");}subClass.prototype=Object.create(superClass&&superClass.prototype,{constructor:{value:subClass,enumerable:false,writable:true,configurable:true}});if(superClass)Object.setPrototypeOf?Object.setPrototypeOf(subClass,superClass):subClass.__proto__=superClass;}var Page1=/*#__PURE__*/function(_React$Component){_inherits(Page1,_React$Component);function Page1(){_classCallCheck(this,Page1);return _possibleConstructorReturn(this,(Page1.__proto__||Object.getPrototypeOf(Page1)).apply(this,arguments));}_createClass(Page1,[{key:"render",value:function render(){console.log('check jquery $',$);return _react.default.createElement("div",null,_react.default.createElement(_multipleExports.A,null));}}]);return Page1;}(_react.default.Component);var Page3=/*#__PURE__*/function(_React$Component2){_inherits(Page3,_React$Component2);function Page3(){_classCallCheck(this,Page3);return _possibleConstructorReturn(this,(Page3.__proto__||Object.getPrototypeOf(Page3)).apply(this,arguments));}_createClass(Page3,[{key:"render",value:function render(){return _react.default.createElement("div",null);}}]);return Page3;}(_react.default.Component);var Page2=/*#__PURE__*/function(_React$Component3){_inherits(Page2,_React$Component3);function Page2(){_classCallCheck(this,Page2);return _possibleConstructorReturn(this,(Page2.__proto__||Object.getPrototypeOf(Page2)).apply(this,arguments));}_createClass(Page2,[{key:"render",value:function render(){return _react.default.createElement("div",null);}}]);return Page2;}(_react.default.Component);var Dash=function Dash(component){return function(routeProps){var DashElement=(0,_react.createElement)(component,_extends({resource:resource},routeProps));return(0,_react.createElement)("div",{},(0,_react.createElement)(Sidebar,_extends({resource:resource},routeProps)),(0,_react.createElement)(component,_extends({resource:resource},routeProps)));};};var Routes=/*#__PURE__*/function(_React$Component4){_inherits(Routes,_React$Component4);function Routes(){_classCallCheck(this,Routes);return _possibleConstructorReturn(this,(Routes.__proto__||Object.getPrototypeOf(Routes)).apply(this,arguments));}_createClass(Routes,[{key:"render",value:function render(){var listReasource={};return _react.default.createElement(_reactRouterDom.BrowserRouter,null,_react.default.createElement(_reactRouterDom.Switch,null,_react.default.createElement(_reactRouterDom.Route,{exact:true,path:"/login",component:Page1}),_react.default.createElement(_reactRouterDom.Route,{exact:true,path:"/user-lise",component:Page2}),_react.default.createElement(_reactRouterDom.Route,{exact:true,path:"/user/:firstName/:lastName/:id",component:Page3})));}}]);return Routes;}(_react.default.Component);exports.default=Routes;
 
 /***/ }),
 /* 30 */
